Other/fix/stop xray on disconnect (#16)
* Fixed stopping of xray.exe

* Added couple of console outputs
Deleted setting the variable to false in "startXray" function

* Deleting README files from destination resources dir

+++ src/electron/copy_tools.action.mjs
@@ -13,7 +13,7 @@ export async function main(srcDirs, destDir) {
 
         const stat = await fs.stat(srcFile);
 
-        if (stat.isFile()) {
+        if (stat.isFile() && !file.endsWith('README.md')) {
           await fs.copyFile(srcFile, destFile);
           console.log(`${srcFile} is copied to ${destFile}`);
         } else if (stat.isDirectory()) {
